refactor: 리뷰 사항 반영
ValidCheck 유틸 기능 필요한 곳으로 이동
StringExtension 기능 ValidCheck
